Centered index category links
Created new CSS class for 'ul' in index which places links in container to center them. Flexboxes could have been used but I have opted to not use it as flexbox implementation would require a rewrite of how the images are scaled which I have deemed not necessary as it would take extra time which could be spent towards more pressing features and functions.