CSS and about us text changes
Made separate 'Index' section for 'categories' class.
Changed newsletter text into using a 'H4' tag. Also the ul the text was wrapped around has been removed.
Justification: all non interactables should use white font to be consistent.
Due to removing the 'il' tags from the styled 'ul' text another suitable already existing style was chosen.
In addition, if you want to add a style do not add it in the html as it is bad practice. If you wanted to add a special CSS for the li in the ul this should have been specified in the about us section of styles.css