Fixed delete and form management
Delete now can delete the last review.
This involved changing the route and delete function on the controller.
The forms do not also cause an error when the last review is deleted
a 'review.div' has been added to make it fit the container width in order to make sure button isn't too small